+++ package.json
@@ -7,6 +7,7 @@
     "firebase": "^7.2.1",
     "leaflet": "^1.5.1",
     "qrcode": "^1.4.4",
+    "rdflib": "^1.0.6",
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
     "react-google-login": "^5.0.5",
