+++ src/components/Login/index.js
@@ -12,7 +12,7 @@ import {
   Icon,
   Divider,
 } from 'semantic-ui-react';
-import { getConsoleOutput } from '@jest/console';
+//import { getConsoleOutput } from '@jest/console';
 
 const Login = ({ history }) => {
   const [state, actions] = useLoginState();
@@ -24,38 +24,6 @@ const Login = ({ history }) => {
 
   if (Auth.isLoggedIn) {
     return <Redirect to={{ pathname: '/shake' }} />;
-  }
-
-  const geoError = function(error) {
-    console.log(error.code);
-  };
-  const geoSuccess = function(pos) {
-    const crd = pos.coords;
-    uploadMetadada(crd)
-  };
-
-  const uploadMetadada = async crd => {
-    const podURL = JSON.parse(localStorage.getItem('authUser')).pod;
-    const $rdf = require('rdflib');
-    const last_latitude = crd.latitude;
-    const last_longitude = crd.longitude;
-    const FOAF = $rdf.Namespace('http://xmlns.com/foaf/0.1/');
-    const store = $rdf.graph();
-    const fetcher = new $rdf.Fetcher(store);
-    await fetcher.load(podURL);
-    const me = store.sym(podURL);
-    store.add(me, FOAF('last_latitude'), new $rdf.Literal(last_latitude), me.doc())
-    console.log("JOJOJO");
-    const updater = new $rdf.UpdateManager(store);
-    const updatePromise = new Promise((resolve) => {
-      const deletions = [];
-      const additions = [$rdf.st(me, FOAF('last_latitude'), new $rdf.Literal(last_latitude), me.doc()),
-      $rdf.st(me, FOAF('last_longitude'), new $rdf.Literal(last_longitude), me.doc())];
-      updater.update(deletions, additions, resolve);
-      console.log(resolve);
-    });
-    await updatePromise;
-    console.log("JEJEJE");
   }
 
   const login = async e => {
@@ -71,12 +39,6 @@ const Login = ({ history }) => {
             if (res.user) {
               Auth.setLoggedIn(true);
               Auth.setUser(res.user);
-              let authUser = JSON.parse(localStorage.getItem('authUser'));
-              if (res.user && res.user.pod != null){
-                console.log("POD FOUND 2!");
-                console.log(authUser.pod);
-                navigator.geolocation.getCurrentPosition(geoSuccess, geoError);
-              }
             }
             history.push('/shake');
           })
@@ -98,13 +60,6 @@ const Login = ({ history }) => {
           .then(res => {
             Auth.setLoggedIn(true);
             Auth.setUser(res.user);
-            let authUser = JSON.parse(localStorage.getItem('authUser'));
-            console.log(res.user);
-            if (res.user && res.user.pod != null){
-              console.log("POD FOUND 3!");
-              console.log(authUser.pod);
-              navigator.geolocation.getCurrentPosition(geoSuccess, geoError);
-            }
             history.push('/shake');
           })
           .catch(({ message }) => {
